Move build-scripts/combine-remediations.py to ssg/_build_remediations.py
